+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ### Changes this version:
 - Changed units for flux offset to A
 - Properly delete ffb rate connections when changing mainclass
+- Added debug toggle option in help menu
 
 ### Changes since v1.9.x:
 - Support for local encoder index